Add signal handler for volume-added
Still figuring out how to capture the signal emitted when things happen
around the OS. So far, I've got the "volume-added" signal to trigger a
function correctly.
When a new volume like cd/dvd is added, a simple notification appears.
This can then be extended to a menu item in the notification center.